+++ lib/msal-common/src/client/AuthorizationCodeClient.ts
@@ -8,13 +8,20 @@ import { AuthorizationUrlRequest } from "../request/AuthorizationUrlRequest";
 import { AuthorizationCodeRequest } from "../request/AuthorizationCodeRequest";
 import { Authority } from "../authority/Authority";
 import { RequestParameterBuilder } from "../server/RequestParameterBuilder";
-import { GrantType } from "../utils/Constants";
+import { GrantType, AADServerParamKeys } from "../utils/Constants";
 import { ClientConfiguration } from "../config/ClientConfiguration";
 import { ServerAuthorizationTokenResponse } from "../server/ServerAuthorizationTokenResponse";
 import { NetworkResponse } from "../network/NetworkManager";
 import { ScopeSet } from "../request/ScopeSet";
 import { ResponseHandler } from "../response/ResponseHandler";
 import { AuthenticationResult } from "../response/AuthenticationResult";
+import { StringUtils } from "../utils/StringUtils";
+import { ClientAuthError } from "../error/ClientAuthError";
+import { UrlString } from "../url/UrlString";
+import { ServerAuthorizationCodeResponse } from "../server/ServerAuthorizationCodeResponse";
+import { AccountEntity } from "../cache/entities/AccountEntity";
+import { EndSessionRequest } from "../request/EndSessionRequest";
+import { ClientConfigurationError } from "../error/ClientConfigurationError";
 
 /**
  * Oauth2.0 Authorization Code client
@@ -37,7 +44,7 @@ export class AuthorizationCodeClient extends BaseClient {
      */
     async getAuthCodeUrl(request: AuthorizationUrlRequest): Promise<string> {
         const queryString = this.createAuthCodeUrlQueryString(request);
-        return `${this.defaultAuthority.authorizationEndpoint}?${queryString}`;
+        return `${this.authority.authorizationEndpoint}?${queryString}`;
     }
 
     /**
@@ -45,11 +52,14 @@ export class AuthorizationCodeClient extends BaseClient {
      * authorization_code_grant
      * @param request
      */
-    async acquireToken(request: AuthorizationCodeRequest): Promise<AuthenticationResult> {
-
+    async acquireToken(request: AuthorizationCodeRequest, cachedNonce?: string, cachedState?: string): Promise<AuthenticationResult> {
         this.logger.info("in acquireToken call");
+        // If no code response is given, we cannot acquire a token.
+        if (!request || StringUtils.isEmpty(request.code)) {
+            throw ClientAuthError.createTokenRequestCannotBeMadeError();
+        }
 
-        const response = await this.executeTokenRequest(this.defaultAuthority, request);
+        const response = await this.executeTokenRequest(this.authority, request);
 
         const responseHandler = new ResponseHandler(
             this.config.authOptions.clientId,
@@ -58,10 +68,59 @@ export class AuthorizationCodeClient extends BaseClient {
             this.logger
         );
 
+        // Validate response. This function throws a server error if an error is returned by the server.
         responseHandler.validateTokenResponse(response.body);
-        const tokenResponse = responseHandler.generateAuthenticationResult(response.body, this.defaultAuthority);
+        const tokenResponse = responseHandler.generateAuthenticationResult(response.body, this.authority, cachedNonce, cachedState);
 
         return tokenResponse;
+    }
+
+    /**
+     * Handles the hash fragment response from public client code request. Returns a code response used by
+     * the client to exchange for a token in acquireToken.
+     * @param hashFragment
+     */
+    handleFragmentResponse(hashFragment: string, cachedState: string): string {
+        // Handle responses.
+        const responseHandler = new ResponseHandler(this.config.authOptions.clientId, this.cacheManager, this.cryptoUtils, this.logger);
+        // Deserialize hash fragment response parameters.
+        const hashUrlString = new UrlString(hashFragment);
+        const serverParams = hashUrlString.getDeserializedHash<ServerAuthorizationCodeResponse>();
+
+        // Get code response
+        responseHandler.validateServerAuthorizationCodeResponse(serverParams, cachedState, this.cryptoUtils);
+        return serverParams.code;
+    }
+
+    /**
+     * Use to log out the current user, and redirect the user to the postLogoutRedirectUri.
+     * Default behaviour is to redirect the user to `window.location.href`.
+     * @param authorityUri
+     */
+    getLogoutUri(logoutRequest: EndSessionRequest): string {
+        // Throw error if logoutRequest is null/undefined
+        if (!logoutRequest) {
+            throw ClientConfigurationError.createEmptyLogoutRequestError();
+        }
+
+        if (logoutRequest.account) {
+            // Clear given account.
+            this.cacheManager.removeAccount(AccountEntity.generateAccountCacheKey(logoutRequest.account));
+        } else {
+            // Clear all accounts and tokens
+            this.cacheManager.clear();
+        }
+
+        // Get postLogoutRedirectUri.
+        const postLogoutUriParam = logoutRequest.postLogoutRedirectUri ? 
+            `?${AADServerParamKeys.POST_LOGOUT_URI}=${encodeURIComponent(logoutRequest.postLogoutRedirectUri)}` : "";
+
+        const correlationIdParam = logoutRequest.correlationId ? 
+            `&${AADServerParamKeys.CLIENT_REQUEST_ID}=${encodeURIComponent(logoutRequest.correlationId)}` : "";
+        
+        // Construct logout URI.
+        const logoutUri = `${this.authority.endSessionEndpoint}${postLogoutUriParam}${correlationIdParam}`;
+        return logoutUri;
     }
 
     /**
@@ -69,9 +128,7 @@ export class AuthorizationCodeClient extends BaseClient {
      * @param authority
      * @param request
      */
-    private async executeTokenRequest(authority: Authority, request: AuthorizationCodeRequest)
-        : Promise<NetworkResponse<ServerAuthorizationTokenResponse>> {
-
+    private async executeTokenRequest(authority: Authority, request: AuthorizationCodeRequest): Promise<NetworkResponse<ServerAuthorizationTokenResponse>> {
         const requestBody = this.createTokenRequestBody(request);
         const headers: Map<string, string> = this.createDefaultTokenRequestHeaders();
 
@@ -138,12 +195,11 @@ export class AuthorizationCodeClient extends BaseClient {
         // add library info parameters
         parameterBuilder.addLibraryInfo(this.config.libraryInfo);
 
+        // add client_info=1
+        parameterBuilder.addClientInfo();
+
         if (request.codeChallenge) {
             parameterBuilder.addCodeChallengeParams(request.codeChallenge, request.codeChallengeMethod);
-        }
-
-        if (request.state) {
-            parameterBuilder.addState(request.state);
         }
 
         if (request.prompt) {
@@ -160,6 +216,10 @@ export class AuthorizationCodeClient extends BaseClient {
 
         if (request.nonce) {
             parameterBuilder.addNonce(request.nonce);
+        }
+
+        if (request.state) {
+            parameterBuilder.addState(request.state);
         }
 
         if (request.claims) {