+++ lib/msal-common/changelog.md
@@ -1,3 +1,14 @@
+# 1.0.0-beta.3
+- Initialization of B2cTrustedHostList (#1646)
+- Unified Cache support (#1444, #1471, #1519, #1520, #1522, #1609, #1622, #1624, #1655, #1680, #1762)
+- `Request` update in msal-common (#1682, #1771)
+- SilentFlow support (#1711)
+- Utilize `Scopeset` across all libraries (#1770)
+- AccountInfo interface (#1789)
+- `state` support in msal-common (#1790)
+- Removal of SPA Client (#1793)
+- EndSessionRequest (#1802)
+
 # 1.0.0-beta.2
 - Fixed an issue where types were not being exported from the correct location (#1613)
 - Fixed an issue where system configuration values were being overwritten with `undefined` (#1631)
@@ -11,7 +22,7 @@
 - Synced all classes and objects to work for both @azure/msal-browser and @azure/msal-node (#1552)
 - Merged configuration for node and browser classes (#1575)
 - Fixed issue with caching for multiple resources (#1553)
-- Adding support for node classes 
+- Adding support for node classes
     - Refresh token client (#1496)
     - Device code client (#1550, #1434)
     - Authorization Code Client (#1434)